+++ sdk/storage/Azure.Storage.DataMovement/tests/PageBlobStorageResourceTests.cs
@@ -158,8 +158,8 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                 await storageResource.CopyFromStreamAsync(
                     stream: stream,
                     streamLength: length,
-                    completeLength: length,
-                    overwrite: false);
+                    overwrite: false,
+                    completeLength: length);
             }
 
             BlobDownloadStreamingResult result = await blobClient.DownloadStreamingAsync();
@@ -192,8 +192,8 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                     stream: stream,
                     streamLength: length,
                     overwrite: false,
-                    position: readPosition,
-                    completeLength: length*2);
+                    completeLength: length * 2,
+                    options: new StorageResourceWriteToOffsetOptions() { Position = readPosition });
             }
 
             BlobDownloadStreamingResult result = await blobClient.DownloadStreamingAsync(
@@ -227,9 +227,10 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                 await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
                 storageResource.CopyFromStreamAsync(
                     stream: stream,
+                    streamLength: length,
                     overwrite: false,
-                    position: position,
-                    streamLength: length),
+                    completeLength: length,
+                    options: new StorageResourceWriteToOffsetOptions() { Position = position }),
                 e =>
                 {
                     Assert.AreEqual(e.ErrorCode, "InvalidHeaderValue");
@@ -581,7 +582,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
 
             // Act
             await TestHelper.AssertExpectedExceptionAsync<RequestFailedException>(
-                destinationResource.CopyBlockFromUriAsync(sourceResource, new HttpRange(0, Constants.KB), false),
+                destinationResource.CopyBlockFromUriAsync(sourceResource, new HttpRange(0, Constants.KB), false, 0),
                 e =>
                 {
                     Assert.AreEqual(e.ErrorCode, "CannotVerifyCopySource");
@@ -650,8 +651,7 @@ namespace Azure.Storage.DataMovement.Blobs.Tests
                     stream: stream,
                     streamLength: length,
                     completeLength: length,
-                    overwrite: false,
-                    position: 0);
+                    overwrite: false);
             }
 
             // Act