+++ sdk/storage/Azure.Storage.Blobs.ChangeFeed/CHANGELOG.md
@@ -1,6 +1,16 @@
 # Release History
 
-## 12.0.0-preview.37 (Unreleased)
+## 12.0.0-preview.38 (Unreleased)
+
+### Features Added
+
+### Breaking Changes
+
+### Bugs Fixed
+
+### Other Changes
+
+## 12.0.0-preview.37 (2023-08-08)
 - Added support for service version 2023-05-03 and 2023-08-03.
 
 ## 12.0.0-preview.36 (2023-07-11)